+++ src/components/TypeSelector.tsx
@@ -31,7 +31,7 @@ export const TypeSelector = (props: { query: InfinityQuery; onChange: (e: Infini
         Type
       </InlineFormLabel>
       <div style={{ marginRight: '5px' }} data-testid="infinity-query-type-selector">
-        <Select width={16} options={getTypes()} onChange={(e) => onTypeChange(e.value as InfinityQueryType)} value={query.type || 'json'} menuShouldPortal={true}></Select>
+        <Select width={30} options={getTypes()} onChange={(e) => onTypeChange(e.value as InfinityQueryType)} value={query.type || 'json'} menuShouldPortal={true}></Select>
       </div>
     </>
   );